refactor: 분리 함수 파일 modules 폴더 생성
