+++ __tests__/CountInputTest.js
@@ -1,4 +1,4 @@
-const countInput = require("../src/countInput.js");
+const countInput = require("../src/modules/countInput.js");
 const App = require("../src/App.js");
 
 describe("스트라이크와 볼 갯수 카운트 기능 테스트", () => {
