+++ src/modules/countInput.js
