+++ src/modules/getComputerNumber.js
