+++ src/modules/isRestartNumber.js
