+++ src/modules/isValidUser.js
