+++ src/modules/printResult.js
